+++ tests/test_files/HIF-non-compliant/bad_edge_field.json
@@ -1,0 +1,9 @@
+{
+    "incidences": [],
+    "edges": [
+        {
+            "edge": 2,
+            "test": "this should break"
+        }
+    ]
+}
