+++ tests/test_files/HIF-non-compliant/bad_edge_without_id.json
@@ -1,0 +1,6 @@
+{
+    "incidences": [],
+    "edges": [
+        {}
+    ]
+}
