+++ tests/test_files/HIF-non-compliant/bad_node_field.json
@@ -1,0 +1,9 @@
+{
+    "incidences": [],
+    "nodes": [
+        {
+            "node": 2,
+            "test": "this should break"
+        }
+    ]
+}
